Site and auxiliary models to new 'sites' django app
