+++ runhaz.sh
@@ -108,7 +108,16 @@ generate_api_schema() {
 
 graph_models() {
     print_style "Generating Entity Relationship Diagram...\n" "success";
-    exec_cmd="$base_py_cmd graph_models trak -g --rankdir=RL --arrow-shape=normal -o $base_dir/docs/haztrak_book/src/assets/erd.png"
+    exec_cmd="$base_py_cmd graph_models sites trak \
+    -g \
+    --pygraphviz \
+    --settings haztrak.settings \
+    --hide-edge-labels \
+    --rankdir=RL \
+    --arrow-shape=normal \
+    -E \
+    --disable-abstract-fields \
+    -o $base_dir/docs/haztrak_book/src/assets/erd.png"
     eval "$exec_cmd"
     exit
 }
