+++ server/apps/sites/models/__init__.py
@@ -1,0 +1,3 @@
+from .contact_models import Address, Contact, SitePhone
+from .epa_profile_models import EpaProfile, SitePermission
+from .site_models import EpaSite, EpaSiteType, Site
