+++ server/apps/sites/models/epa_profile_models.py
@@ -1,10 +1,64 @@
+from django.contrib.auth.models import User
 from django.core.exceptions import ValidationError
 from django.db import models
 
-from apps.sites.models import Site
+from .base_models import SitesBaseModel
+from .site_models import Site
 
-from .base_model import TrakBaseModel
-from .rcra_profile_model import RcraProfile
+
+class EpaProfile(SitesBaseModel):
+    """
+    Provides the user's EpaProfile information, excluding their RCRAInfo
+    API key (see ProfileUpdateSerializer). Has a one-to-one relationship with
+    the User model.
+    """
+
+    class Meta:
+        ordering = ["rcra_username"]
+
+    user = models.OneToOneField(
+        User,
+        on_delete=models.CASCADE,
+    )
+    rcra_api_key = models.CharField(
+        max_length=128,
+        null=True,
+        blank=True,
+    )
+    rcra_api_id = models.CharField(
+        max_length=128,
+        null=True,
+        blank=True,
+    )
+    rcra_username = models.CharField(
+        max_length=128,
+        null=True,
+        blank=True,
+    )
+    phone_number = models.CharField(
+        max_length=15,
+        null=True,
+        blank=True,
+    )
+    email = models.EmailField()
+
+    def __str__(self):
+        return f"{self.user.username}"
+
+    def sync(self):
+        """Launch task to sync use profile. ToDo: remove this method"""
+        from apps.trak.tasks import sync_user_sites
+
+        task = sync_user_sites.delay(str(self.user.username))
+        return task
+
+    @property
+    def is_api_user(self) -> bool:
+        """Returns true if the use has Rcrainfo API credentials"""
+        if self.rcra_username and self.rcra_api_id and self.rcra_api_key:
+            return True
+        return False
+
 
 EPA_PERMISSION_LEVEL = [
     ("Certifier", "Certifier"),
@@ -13,9 +67,9 @@ EPA_PERMISSION_LEVEL = [
 ]
 
 
-class SitePermission(TrakBaseModel):
+class SitePermission(SitesBaseModel):
     """
-    RCRAInfo Site Permissions per module connected to a user's RcraProfile
+    RCRAInfo Site Permissions per module connected to a user's EpaProfile
     and the corresponding Site
     """
 
@@ -25,7 +79,7 @@ class SitePermission(TrakBaseModel):
 
     site = models.ForeignKey(Site, on_delete=models.CASCADE)
     profile = models.ForeignKey(
-        RcraProfile,
+        EpaProfile,
         on_delete=models.PROTECT,
         related_name="site_permission",
     )
