+++ server/apps/trak/models/__init__.py
@@ -1,13 +1,11 @@
 """
 Model definitions for the trak domain
 """
-from .address_model import Address
+
 from .code_models import WasteCode
-from .contact_model import Contact, EpaPhone
-from .handler_model import EpaSite, ManifestHandler
-from .manifest_model import Manifest
-from .rcra_profile_model import RcraProfile
-from .signature_model import ESignature, PaperSignature, QuickerSign, Signer
-from .site_permission_model import SitePermission
-from .transporter_model import Transporter
-from .waste_model import WasteLine
+from .contact_models import EpaPhone
+from .handler_models import ManifestHandler
+from .manifest_models import Manifest
+from .signature_models import ESignature, PaperSignature, QuickerSign, Signer
+from .transporter_models import Transporter
+from .waste_models import WasteLine
