+++ server/apps/trak/models/code_models.py
@@ -3,7 +3,7 @@ import logging
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from apps.trak.models.base_model import TrakBaseModel
+from apps.trak.models.base_models import TrakBaseModel
 
 logger = logging.getLogger(__name__)
