+++ server/apps/trak/services/site_service.py
@@ -3,9 +3,8 @@ from typing import Dict, List
 
 from django.db import transaction
 
-from apps.sites.models import Site
+from apps.sites.models import EpaSite, Site
 
-from ..models import EpaSite
 from .manifest_service import ManifestService
 from .rcrainfo_service import RcrainfoService
