+++ server/apps/trak/signals.py
@@ -2,13 +2,13 @@ from django.contrib.auth.models import User
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
-from apps.trak.models import RcraProfile
+from apps.sites.models.epa_profile_models import EpaProfile
 
 
 @receiver(post_save, sender=User)
 def create_profile(sender, instance, created, **kwargs):
     if created:
-        RcraProfile.objects.create(user=instance)
+        EpaProfile.objects.create(user=instance)
 
 
 @receiver(post_save, sender=User)
